Applied patch from issue 12, reported by vinvin. Two calls to http_MakeMessage were broken from large file support introduced in r79
git-svn-id: https://pupnp.svn.sourceforge.net/svnroot/pupnp/trunk@86 119443c7-1b9e-41f8-b6fc-b9c35fce742c

+++ upnp/src/soap/soap_ctrlpt.c
@@ -650,7 +650,8 @@ SoapSendAction( IN char *action_url,
 
     // make request msg
     request.size_inc = 50;
-    if( http_MakeMessage( &request, 1, 1, "q" "N" "s" "sssbs" "U" "c" "bbb", SOAPMETHOD_POST, &url, xml_start_len + action_str_len + xml_end_len,   // content-length
+    if( http_MakeMessage( &request, 1, 1, "q" "N" "s" "sssbs" "U" "c" "bbb", SOAPMETHOD_POST, &url, 
+                          (off_t)xml_start_len + action_str_len + xml_end_len,   // content-length
                           ContentTypeHeader,
                           "SOAPACTION: \"", service_type, "#", name.buf,
                           name.length, "\"\r\n", xml_start, xml_start_len,
